test: backdate fixture root so itemize plans are deterministic
The dry-run and itemize checks pulled a fixture whose entries were
backdated but whose source *root* directory kept a "now" mtime. rsync
itemizes the transfer root as `.d..t...... ./` only when the source root
mtime differs from the destination root, and the destination is freshly
created (mtime "now") right before each pull. So whether the `./` row
appeared depended on whether the source-build and each dest-creation
landed in the same clock second - a race. Running upstream then oc
sequentially, the two could straddle a second boundary differently,
making their itemize plans compare unequal and the check fail
intermittently across transports. Verified on Linux: with the root
mtime distinct, upstream and oc emit byte-identical rows (the `./` row
present on both); when it matches, neither emits it.

Backdate the source root to the same fixed past mtime as the entries so
the `./` row is deterministic (always present, identical) for every
client. The dry-run and itemize fixtures were byte-identical, so fold
them into one shared `support::build_backdated_tree` helper - fixing the
flake in one place and removing the duplication.
